Change volume target from float to int

+++ background_scripts/storage.js
@@ -5,7 +5,7 @@ function sendVolumeTargetOnStartup() {
         localStorage.setItem("volumeTarget", "-14");
         volumeTarget = -14;
     } else {
-        volumeTarget = parseFloat(volumeTarget);
+        volumeTarget = parseInt(volumeTarget);
     }
 
     browser.tabs.query({}).then((tabs) => {
@@ -18,11 +18,11 @@ function sendVolumeTargetOnStartup() {
 // Send volume target when popup is opened, store volume target when changed in popup
 browser.runtime.onMessage.addListener((newValues) => {
     if (newValues === "popupOpen") {
-        browser.runtime.sendMessage({volumeTarget: parseFloat(localStorage.getItem("volumeTarget"))});
+        browser.runtime.sendMessage({volumeTarget: parseInt(localStorage.getItem("volumeTarget"))});
     } else if (newValues === "injected") {
         browser.tabs.query({}).then((tabs) => {
             for (const tab of tabs) {
-                browser.tabs.sendMessage(tab.id, parseFloat(localStorage.getItem("volumeTarget")));
+                browser.tabs.sendMessage(tab.id, parseInt(localStorage.getItem("volumeTarget")));
             }
         });
     } else {
